feedem-v2/requirements.txt
pyyaml==5.4.1
paho-mqtt==1.6.1